Exercise 10

+++ 10-binary_tree_depth.c
@@ -8,9 +8,9 @@
  * Return: Depth of the node
  */
 
-int binary_tree_depth(const binary_tree_t *tree)
+size_t binary_tree_depth(const binary_tree_t *tree)
 {
-	int depth = 0;
+	size_t depth = 0;
 
 	if (tree == NULL)
 		return (0);
